backend/pyproject.toml
[project]
name = "smartfiles"
version = "0.1.0"
description = "AI-powered semantic search for document folders"
dependencies = [
    "typer",
    "fastapi",
    "uvicorn",
    "sentence-transformers",
    "chromadb",
    "pypdf",
    "pillow",
    "pytesseract"
]

[project.scripts]
smartfiles = "smartfiles.cli.main:app"